Enforce active turn for first warring human
Who knows if this will have bad side-effects, but it seems to
work in the lobby.

Test plan:

Load up my bad save game and check that India's turn is active
in the lobby. It does. Hopefully we can make it through India's
turn and then through every other warring players turns too.
Fingers crossed on that.

Also, I tried to reproduce the bad save game in a two player
game. Unfortunately (?), this game seems to load fine. I can
load this working save with the patch applied, and have verified
that it doesn't break anything.